finance_mvp/backend/requirements.txt
fastapi==0.115.9
uvicorn[standard]==0.34.0
python-multipart==0.0.20
sqlalchemy==2.0.39
psycopg[binary]==3.2.6
pydantic==2.10.6
pydantic-settings==2.8.1
alembic==1.14.1
pandas==2.2.3
pdfplumber==0.11.5
pytesseract==0.3.13
Pillow==11.1.0
rapidfuzz==3.12.2
python-dateutil==2.9.0.post0
scikit-learn==1.6.1
numpy==2.2.3
